+++ packages/sitecore-jss/src/editing/graphql-editing-service.test.ts
@@ -15,6 +15,7 @@ import {
   mockEditingServiceResponse,
 } from '../test-data/mockEditingServiceResponse';
 import { EditMode } from '../layout';
+import { LayoutKind } from './models';
 import debug from '../debug';
 
 use(spies);
@@ -91,12 +92,20 @@ describe('GraphQLEditingService', () => {
       })
     ).to.be.true;
     expect(clientFactorySpy.returnValues[0].request).to.be.called.exactly(1);
-    expect(clientFactorySpy.returnValues[0].request).to.be.called.with(query, {
-      language,
-      version,
-      itemId,
-      siteName,
-    });
+    expect(clientFactorySpy.returnValues[0].request).to.be.called.with(
+      query,
+      {
+        language,
+        version,
+        itemId,
+        siteName,
+      },
+      {
+        headers: {
+          sc_layoutKind: 'final',
+        },
+      }
+    );
 
     expect(result).to.deep.equal({
       layoutData: layoutDataResponse,
@@ -148,12 +157,20 @@ describe('GraphQLEditingService', () => {
       })
     ).to.be.true;
     expect(clientFactorySpy.returnValues[0].request).to.be.called.exactly(1);
-    expect(clientFactorySpy.returnValues[0].request).to.be.called.with(query, {
-      language,
-      version,
-      itemId,
-      siteName,
-    });
+    expect(clientFactorySpy.returnValues[0].request).to.be.called.with(
+      query,
+      {
+        language,
+        version,
+        itemId,
+        siteName,
+      },
+      {
+        headers: {
+          sc_layoutKind: 'final',
+        },
+      }
+    );
 
     expect(result).to.deep.equal({
       layoutData: {
@@ -203,6 +220,55 @@ describe('GraphQLEditingService', () => {
       siteName,
       version: undefined,
     });
+
+    expect(result).to.deep.equal({
+      layoutData: layoutDataResponse,
+      dictionary: {
+        foo: 'foo-phrase',
+        bar: 'bar-phrase',
+      },
+    });
+
+    spy.restore(clientFactorySpy);
+  });
+
+  it('should fetch shared layout editing data', async () => {
+    nock(hostname, { reqheaders: { sc_editMode: 'true', sc_layoutKind: 'shared' } })
+      .post(endpointPath, /EditingQuery/gi)
+      .reply(200, editingData);
+
+    const clientFactorySpy = sinon.spy(clientFactory);
+
+    const service = new GraphQLEditingService({
+      clientFactory: clientFactorySpy,
+    });
+
+    spy.on(clientFactorySpy.returnValues[0], 'request');
+
+    const result = await service.fetchEditingData({
+      language,
+      version,
+      itemId,
+      siteName,
+      layoutKind: LayoutKind.Shared,
+    });
+
+    expect(clientFactorySpy.calledOnce).to.be.true;
+    expect(clientFactorySpy.returnValues[0].request).to.be.called.exactly(1);
+    expect(clientFactorySpy.returnValues[0].request).to.be.called.with(
+      query,
+      {
+        language,
+        version,
+        itemId,
+        siteName,
+      },
+      {
+        headers: {
+          sc_layoutKind: 'shared',
+        },
+      }
+    );
 
     expect(result).to.deep.equal({
       layoutData: layoutDataResponse,